Update condition for storing chat options
Fixes issue where, when using server-side threads, AF would pass back the function call message due to store being in chat_options in kwargs, not at the kwargs level. Means the appending of only the tool execution result never happens.

Resulted in a "tool output not found error" as we replied with the call not the result first.